improve tab handling
