+++ digitransit-component/packages/digitransit-component-autosuggest/src/utils/utils.js
@@ -6,6 +6,18 @@ import { getStopName } from '@digitransit-search-util/digitransit-search-util-he
 import getLabel from '@digitransit-search-util/digitransit-search-util-get-label';
 import { DateTime } from 'luxon';
 
+export const isKeyboardSelectionEvent = event => {
+  const space = [13, ' ', 'Spacebar'];
+  const enter = [32, 'Enter'];
+  const key = (event && (event.key || event.which || event.keyCode)) || '';
+
+  if (!key || !space.concat(enter).includes(key)) {
+    return false;
+  }
+  event.preventDefault();
+  return true;
+};
+
 export const isOriginDestinationOrViapoint = id =>
   id === 'origin' ||
   id === 'destination' ||
